Change locking implementation works
Needed to add items found in previous cache to current cache.

When an item is not found in the current cache, but is found in the previous cache, that item needs to be put into the current cache.  Doing this requires a write lock on the cache. In odrer to avoid locking the entire cache, instead just acquire a lock for any key that would affect the same part of the tree.

Put needed to check for existing values in the previous cache.

Also added unit test for cache rotation.

There was a need to