Discover provider using lotus (#38)
* Discover provider using lotus

This looks up a miner using the miner short address (e.g. ”t01000") and gets the miner's peerID and addresses.  If the requested peer id matches the one returned and the signature in the request is verified, the miner is consider discovered and my participate in indexing.

* fix go.mod and unit tests
* No need to support lotus as local node
* rebase on main and resolve conflicts
* Stop leaving behind temp dirs

Note: merging despite test failures due to need for later version of lotus/IPLD/graphsync which are not available yet.