Add a provider client
storetheindex should provide a provider-client that is used to communicate with providers.  storetheindex will use this client, and if someone wants to write their own indexer, they are welcome to also use this provider-client.

Simplified the libp2pclient so that there a separate client instance is created for each different peer to communicate with.  This means that there is no need to maintain and synchronize a map of different peers within a single client.  THis is advantageous when the app or goroutine only connects to a single peer.