api/v0: reject Metadata.ProtocolID == 0
Integers in Go default to the zero value, meaning that if the user
forgets to set the protocol ID, it could transparently default to a
confusing value. For instance:

	metadata := v0.Metadata{Data: someData}

This could lead to confusing decode errors later on, or to metadata
being mistaken for the wrong kind.

Instead, reject this zero value. The protocol ID is a multihash anyway,
and we want users to either select registered indexer metadata
multihashes, or a multihash in the reserved range for testing.

The check happens at both encode and decode, for extra safety.
This required adding an error return on the encode side.
While we're there, make Metadata implement encoding's Binary interfaces.

Updates #94.