Upgrade to latest `go-legs`
Upgrade dependency to `go-legs`. The upgrade introduces a breaking
change where a subscriber needs to also specify a selector. The selector
specified aims to recursively sync the graph of advertisements and their
chain of entry chunks. This is equivalent to how the indexer node
operated before selectors existed in legs.

Relates to:
 - https://github.com/filecoin-project/indexer-reference-provider/pull/81