Update to new version of go-legs
The new go-legs works significantly differently than the previous version.  It does not require explicitly subscribing to peers from which to receive advertisements.  Instead, this is now controlled by the configured discovery policy: advertisements are read and state is kept for any providers that are authorized.  This also means that the Subscribe and Unscribe admin APIs no longer exist.  Instead there are now Allow and Block to accept or reject provider advertisements.

Syncing can be done to an http provider if there is an explicit sync to an HTTP multiaddress.  The sync command now has an optional flat to specify the multiaddr of the provider to sync to.  This is used for an HTTP sync or a sync where the address is not already known to the indexer through discovery.