Update to new version of go-legs (#143)
* Update to new version of go-legs

The new go-legs works significantly differently than the previous version.  It does not require explicitly subscribing to peers from which to receive advertisements.  Instead, this is now controlled by the configured discovery policy: advertisements are read and state is kept for any providers that are authorized.  This also means that the Subscribe and Unscribe admin APIs no longer exist.  Instead, there are now Allow and Block APIs to accept or reject provider advertisements.

Syncing can be done to an http provider if there is an explicit sync to an HTTP multiaddress.  The sync command now has an optional flag to specify the multiaddr of the provider to sync to.  This is used for an HTTP sync or a sync where the address is not already known to the indexer through discovery.

* New command to reload policy from config file
* Depend on go-legs v0.2.1
* Updated comments and logging
* Sync remove advertisements in order, handle remove by context ID ads
* Fix race on Close
* Allow publisher and provider to be different peers
* Update comments to clarify differences between publisher and provider.
* Allow/trust policy pertains to both publishers and providers.
* Verifying the advertisement signature returns the signer's peer ID so that it can be checked if it is allowed.
* Check that the advertisement is signed by the provider specified in the advertisement.
* Update tests to check that indexed data uses provider ID from advertisement.
* Update tests to sign advertisement using private key that matches provider ID.
* Do not sync ad that is already being synced