Support advertisement to update metadata only
This PR also makes it possible to support advertisements for the removal of content by context ID, or removal of individual entries

A special cidLink value, schema.NoEntries, is used to indicate that an advertisement has no entries.  When this is used in a removal advertisement, there are no specific indexes to remove and all content with the matching context for the provider is removed.  If a removal advertisement has an entries link to actual entries, then the entries are retrieved from the provider and the individual multihashes within each entry are used to remove content.

If an advertisement uses the special schema.NoEntries value as its entry link, and the ad is not a not for removal, then the advertisement is used only to update the provider's metadata that has the matching context ID.