Process content entries in earliest to latest order (#166)
* Process content entries in earliest-to-latest order

When handling a chain of entries, it was still possible for content entries to be processed out of order if a previous ad took too long to show up.  This adds a timer to wait for a previous ad if there is one that has not already been handled by the indexer.