Mark an advertisement as processes when the is has no entries
When processing an ad with no entries (Entries set to NoEntries), the ad was not being marked as processed.  This fixes that so that those ads are marked as processed. If the ad is not readable then it will also be marked as processes so that the indexer does not get stuck trying to reprocess the same unreadable ad.

- Update go-legs and libp2p