Mark an advertisement as processed when the is has no entries (#190)
* Mark an advertisement as processed when the is has no entries

When processing an ad with no entries (Entries set to NoEntries), the ad was not being marked as processed.  This fixes that so that those ads are marked as processed. If the ad is not readable then it will also be marked as processed so that the indexer does not get stuck trying to reprocess the same unreadable ad.

* Update go-legs, go-libp2p, go-indexer-core
* Add test for no entries