New Ingester System (#224)
* Notify when new ad ingested completely

Add the ability to receive notification when a new advertisement and its entries have been ingested.

* Stabilize tests and fix wait for previous advertisement

There was an error in waiting for the previous ad, which resulted from moving some code from the StorageWriteOpener into the StorageHook.  When waiting for the previous ad, if that ad was not yet wait-able, the code would first check to see if the ad was in the datastore to determine if it had already been processed.  This was unreliable because the ad would already be in the datastore whether or not is was already processed.  This was because it was put into the datastore in the Storage write opener, but this check was done in the StorageHook.

This is fixed now by watching for the previous block to be processed for a time, and if it does not get processed during that wait, then check for wait-ability once more and wither wait or move on because the block in not being synced.

The Ingester.Sync() method returns a read-only channel of cid.Cid that the caller can optionally wait on.  Nothing is put onto the channel until Sync has completed syncing up to the head advertisement.

- This PR includes an unrelated fix in the client to close hosts it creates.
- The testEnv type now has a Close method to clean up resources.

* Fix a couple issues, and skip a test

- Close notice channels on close
- Always signal ad sync done
- Skip Restart test

* Fix test

* Use memtransport

* Faster mkTestHost

* Tweaks to memtransport

* Faster TestMultiplePublishers

* Fix import

* Faster TestSubscribe

* Faster TestRestartDuringSync

* Skip e2e test

* Revert decreasing waitPrevAdTime. Try increasing

* Test not using cidwaiter

* Skip memtransport test

* Skip TestRegisterProvider http handler

* Remove memtransport file

* WIP – new ingest system

* Fix race in pending syncs

* Remove cidwaiter

* Remove push/pop chunk<->cid mapping

* Cleanup

* Add all work to the queue

* Cleanup perf logs in test

* Check if chan was closed

* Use default Ingest config in tests

* Fix race

* mod tidy

* Add lint line

* Fix comment

* Simplify staging logic

* Remove duplicate close call

* Rename

* Fix comment

* For range over channel

* Bump go-legs and fix rebase errors

* Test that manual sync with depth works

* Populate unset IngestWorkerCount in config

* Add log message at start of ad processing

* fix comment

* Allow for reprocessing of ads

* Remove skipKey logic

* Add test around failure during resync

* Minimize error stack in tests

* Use defer to always delete entry chunk

* Fix tense

Co-authored-by: gammazero <gammazero@users.noreply.github.com>