Return more information on ad ingest errors
Now the caller can decide when to bail early or when to skip processing
the ad. This gets us really close to being able to pass errors to the
instigator.

Also fixes a bug in the logic of skipping processing earlier ads if a
later ad has already been processed.