Indexer auto-syncs after period of no contact with publisher
If the indexer has had no contact with any publisher for a provider, the indexer attempts to sync with the provider's last known publisher.  As long as there is no contact with a publisher for the provider, the indexer continues retrying to sync periodically until contact is made or until the indexer gives up after a configurable abount of time.