Don't optimisitically delete ad when loadingAd
This fixes a bug where an ingest loop loads an ad (and thereby deleting
it) and for whatever reason fails to finish ingesting the ad. The next
time the loop tries to ingest the ad it will fail to load the ad and
skip it.

We don't need to delete the ad here, it will be deleted when we mark the
ad processed. Then it's fine if some future ingest loop fails to load
the ad since we already processed it anyways.