Remove recursion depth limits for sync
Limiting traversal depth for advertisement chains and entry chains is not a reliable mechanism for preventing DoS, and can cause the indexer to serve incorrect content if advertisements are missed, and incomplete content if entries are missed when the limits are configured too low.  Rate limiting will be used to prevent busy publishers from starving others, instead of depth limiting.