Configure EBS CSI controller to run on all nodes
The EBS CSI controller installed by EKS as addon doesn't seem to provide
taint control configuration. Install the controller directly so that:
1. we have better control over its configuration.
2. we avoid clash with CD-managed manifests applied via gitops

With this change, the controller daemonset is configured to run on all
nodes regardless of their taint. Before this change, the daemonset was
not being scheduled on r5b nodes which meant we could not create io2
volumes on those nodes which defeated the whole point of having r5b nod
types.

The changes here also create the IAM role needed by the controller to do
its job and patches the corresponding controller service account. This
 was missing in earlier PRs.