Create GitHub Actions credentials and set container publish to ECR
Create an IAM role with minimal permissions for GitHub Actions that
allows it to publish images to public ECR repositories owned by
storetheindex. The role uses GitHub OIDC provider for authentication and
is only allowed to be assumed for builds in storetheindex repo.

Set up a GitHub Actions build that is triggered on either merge to main
branch or tag `v*` that builds and publishes container images to the
storetheindex public ECR repo. The container image tag is determined
based on the event that starts up the build:
- if the build is triggered by a tag starting with `v` then the
container image will be whatever came after the `v` in tag, i.e. semver.
- otherwise, it will be a timestamp at which build was run plues `-`
plus the build commit SHA. Note that to configure automatic CD
deployment we need some sortable numeric or alphabetical value and using
commit SHA alone will not suffice; hence the timestamp prefix.