Spread indexer instances across AZs when possible
To improve resilience against AZ failure, we want the indexer instances
to be spread across multiple AZs. So that, failure of one AZ does not
result in downtime.

Increase worker count of type r5b in prod so that e have one in each AZ.

Configure topology spread constraints for indexer statefulsets in dev
and prod such that they are _preferably_ spread across the AZ in which
each worker node runs.