If no value for "indexer" CLI flag, then read from config (#407)
* If no value for "indexer" CLI flag, then read from config

This removes the need to specify the `indexer` CLI flag when running commands against a local indexer.

Usually the indexer listens on a port other than 80, but if the `indexer` flag is not specified for CLI commands, then localhost:80 is used as the default.  This PR changes CLI the behavior so that when no `indexer` flag is given, the value is read from the config file if available.  Depending on the command, either the finder address or the admin address is read.

* Default to localhost if not address available