Direct announce messages use go-legs Announce
This allows any number of incoming announce requests while the indexer is busy syncing the peer that the announce messages are for, without building up calls to Sync or extra goroutines for each announce.

This also includes a fix to go-legs to prevent goroutines building up from pubsub messages as well.