Onboard `autoretrieve` as a temporary tenant of sti `dev` cluster
While `autoretrieve` is being developed, provide a temporary place it
can be deployed in such that metrics could be forwarded to PL grafana
along with logs etc. All of this is already set up in sti EKS clusters.

Create a dedicated ECR repository for automating the build and push of
`autoretrieve` container images. It looks like the exising one is
manually created and pushed to. Here we create a separate one to avoid
any conflicts.