GitOps deployment of dev-v2 cluster (#449)
Deploy storetheindex using the self managed manifests on a cluster called dev-v2 that is created using CAPI, and AWS resources created using crossplane managed by WeaveWorks enterprise via:

- https://github.com/filecoin-project/gitops-root/blob/main/clusters/wge-prod-capi-configs/sti-dev-v2.yaml

Note that the "v2" signifies a second iteration of gitops by weave works where the set-up is very similar to what storetheindex has already. The point of integration with dev is to learn what it takes to evolve a v2 (like the one we have) to the planned v3 that will materialise in the future.