Allow `put` via cloudfront sitting in front of `cid.contact`
`cid.contact` should allow `PUT` to `/ingest/announce` for go-legs HTTP
publishers. Further, `cid.contact` is fronted by the cloudfront caching
and `PUT` was not allowed in the cloudfront behaviour, which was causing
`announce` requests from go-legs HTTP publishers to fail with `403`.

Allow all methods on cloudfront caching, because we need `GET` and `PUT`
to be allowed but they can't be configured explicitly. CloudFront only
supports `GET, HEAD` or` GET, HEAD, OPTIONS` or
`GET, HEAD, OPTIONS, PUT, POST, PATCH, DELETE` hence the exhaustive
allowed methods list.

Same is applied to `dev` while we are at it.