Use larger instance types and override config in dev to limit providers
The dev instances are still running on 10GiB memory because the maximum
available on the worker node type is 16. It also uses a configuration
identical to prod, which means it would ingest everything prod does.

Here we whitelist only a selected number of providers to reduce load on
the `dev` instance since it is meant to be for development purposes.
This should shield it from failures that say prod instance goes through
due to excessive ingest load.

The memory given to the instances as well as the instance types are
upgraded to match production. Taint is set on the dev instance to use
r5b instance types just like prod. This will allow us to change the dev
config as we see fit to match the prod config or selectively ingest from
specific providers for debuging purposes without having to worry about
mismatching hardware requirements.