Fix intermittent test failures
In `TestEndToEndWithReferenceProvider` reset `Stdin` and `Stderr` for
the command before running it in a combined output. Otherwise,
`CombinedOutput` will return an error. Increase total E2E test timeout
to 2 minutes, since it seems to intermittently fail on MacOS.

In `TestMultiplePublishers`, assert indices from each test publisher are
eventually indexed since the background ad processing might not have
finished when the sync returns. Assert Synced CID is equal to the
expected head before checking reference to the latest synced.

In `TestPollProvider` registry tests close the registry to avoid failure
in Windows when tests attempt to remove temporary files that otherwise
will be in use and cannot be deleted.