Improve test stability by asserting multihashes are eventually indexed
Across ingest tests assert that target multihashes are eventually
indexed. Because, there is no guarantee that they are indexed after an
explicit sync returns. This should improve stability and reproducibility
of the tests.

Introduce test helper method to select entries only from a singe ad
rather than a given ad and all its previous ads. This helps make the
tests more concise in that, the assertions would precisely assert which
multihashes should not have been indexed, instead of asserting partial
indexing of multihashes.

Refactor tests to use existing utility functions with less repeated
logic when it comes to asserting that multihashes are not indexed.