Fix bug where tail of chain is synced twice
Fix an issue where the tail of a chain, either ad or entry chunk is
resynced. Block hook was never calling the actions if the next link was
nil, meaning we have reached the end of the chain.

As a result the CID before last was being left as the CID to try in the
next segmented sync and therefore the segmented sync cycle continued.

The fix explicitly sets the next CID to `Undef` when we have reached the
end of a chain which then signals that segmented sync cycle must stop.

Upgrade to the latest go-legs.