Increase CPU resources in `prod` to `3.5`
The CPU usage in `prod` is at limit and the advertisement processing
is relatively slow. The worker nodes running the `prod` instances have
4 compute capacity and we can increase the CPU limit for indexers to
3.5, leaving 0.5 for the cluster related pods.

The `dev` environment is using the same worker type as prod with the
same cluster set up. Earlier commits confirmed that 3.5 CPU was
successfully scheduled by the K8S cluster.