Defer handling of announce until ad chain processing is finished
In a case where a provider has a consistently growing ad chain where
each ad is explicitly announced while the indexer is in the progress of
processing previously announced ads a race-condition could occur where
the ingestion logic competes with the announce logic for syncing CIDs.
The former competes for syncing the ad entries so that they can be
processed, while the latter competes for syncing the newly announced
ads. As a result, the processing of ads slows down and CPU usage
increases significantly.

To avoid this postpone processing of any explicit announce request while
an ad chain is being processed. Once the ad chain processing finished,
then handle the announce request. Note that we only need to handle the
latest announce request because of the chaining of advertisements; as
long as we know the latest head we can discover the rest.

Relates to: #523