Do not trigger rolling deployment when indexer config changes
The indexer config is mounted as a file via K8S ConfigMap that is
generated in kustomization.yaml. The generated ConfigMaps will include a
hash of their content so that it triggers a rolling deployment on the
pods that mount them. This is a typical set up in a case where
redeployment is fast.

In the case of storetheindex, redeployments are not fast; it takes
around 60-90 minutes for each pod to be ready because on start up the
indices need to be re-scanned.

To reduce the wait time when changing configuration prior work
introduced automatic reload of config when the configuration changes.
The changes here make the name of the configmap fixed so that we can
change the mounted files without having to restart the pods and the
application-level config reload will take care of the rest of config
changes at runtime by watching the config file for change.