Add `INDEXER_WATCH_CONFIG` env var back to avoid race condition
In dev the K8S manifests can get deployed before the container with the
new env var name is built, pushed and deployed. To avoid this race
condition keep the env var name backward-compatible until we are sure
that all the currently deployed containers support the new name.

Note that in prod the containers will not have  this change until we
publish a release which then results in publishing a container that will
include the env var change.