Log levels are configurable in config file and reloadable (#564)
* Log levels are configurable in config file and reloadable

Log levels are now configurable in the config file. An overall log level setting is configurable, as well as the settings for any logger that should not have the same level as the overall setting. These config settings are reloaded when the config file is reloaded at runtime.

If there is no logging configuration, the default log level is set to "info" and the loggers "basichost", "bootstrap", "dt_graphsync", "dt-impl", and "graphsync" are set to level "warn".

* Remove cli flag to override log level, since chaing log level should be done by reloading config