Ingest Entries of type IPLD HAMT
Expand the indexer protocol to accept and ingest `Entries` link in
advertisements that point to IPLD HAMT ADL in addition to the existing
interlinked chain of `EntryChunk`s. This is to facilitate a
backward-compatible ingestion logic that not only can process the
existing chains but also allows the entries link kind to differ across
advertisements in a single chain.

The entries processing logic is modified to automatically detect the
kind of entries and process them accordingly. The logic:
 1. Syncs the root node to which the `Entries` link in an advertisement
    points.
 2. Checks the kind of the synced node.
 3. If HAMT:
  a. Recursively syncs any remaining links in the HAMT.
  b. Batches the keys in the hamt in `Ingest.StoreBatchSize` sizes.
  c. Sends each batch to the existing indexing logic.
  d. Deletes the HAMT from the local datastore.
 4. If `EntryChunk`:
  a. Processes the root node that is synced already as before.
  b. recursively syncs through remaining interlinked `EntryChunk`s as
     before.

The maximum entries recursion depth is modified to 1 plus the configured
max depth. This is because the root entries link is synced separately to
probe the kind of entries we are dealing with before syncing the
remaining chain. A `TODO` is left in code to investigate whether it's
worth modifying the selector on the fly and reducing the recursion depth
by one so that the maximum configured value is respected.

The segmented sync for HAMT entries is disabled and postponed to future
work in order to reduce the scope of this change. A `TODO` is left to
investigate how useful the segmented sync would be for a HAMT before
implementing the logic.

Testing utilities are introduced to generate ad chains with HAMT
entries. All existing tests are modified to use a mix of HAMT and
`EntryChunk` chains for testing purposes.

The IPLD selectors used by the ingester are refactored into a struct
variable better re-usability/readability.

Note that the dependencies use specific versions of the latest
go-data-transfer, go-ipld-prime, and go-ipld-adh-hamt; none of which
have been released yet. It is not possible to use the latest released
versions due to the need for bug fixes.

Fixes: #420