Configure peering service that keeps peers connected (#588)
* Configure peering service that keeps peers connected

Indexers may want to remain connected to other peers for the purpose of receiving pubsub messages. This is different than bootstrapping because the indexer maintains a connection to all peers configured for the peering service, whereas bootstrapping only maintains a connection to the minimum set of bootstrap peers.

The peering configuration is reloadable at runtime. If there are no peers configured for the peering service, then it does not run.

Fixes issue #583

* Update configs and doc
* Update deps