Fix response headers and CORS OPTIONS support via middleware
The finder server responses did not include the headers despite them
being set as part of request handling. Instead of setting headers in
handlers, which may not get applied depending on gorilla mux router
configuration, use a "middleware" provided by the gorilla mux framework
to set allowed methods extracted from the router as well as allow
origin all.

For CORS to function correctly, the endpoints also need to handle
`OPTIONS` method gracefully. This is also added, where the handling is
terminated after the headers are written.

Relates to:
 - https://github.com/filecoin-project/storetheindex/issues/612