Deploy a debug instance of sti in `dev` with 32-bit sth bucket size
Deploy fresh instances of storetheinex in `dev` under a separate
namespace with empty disk and 32-bit bucket size for their underlying
datastore with 60Gi memory and 5 CPUs each.

Note that automatic image update via CD pipeline is not enabled for the
namespace. The rationale is for developers to explicitly update the
image they want to deploy. The image tag is currently set to the head
of `main`, which introduces `STHBits` config value.