Deploy individually configurable sti nodes in `dev`
The current deployment of storetheindex uses a statefulset with replica
count of 2. This makes up a simpler deployment mechanics but does not
allow the config or version of the container run by each node to be
different from one another. This makes debugging or trial of a specific
solution require an "all or nothing" change to the environment.

A new kustomization overlay is defined, which is explicitly designed to
run a single instance of storetheindex. The overlay also separates the
PVC definitions which makes expansion of volumes easier, and removes the
need for manual statefulset re-creation to work around volume claims
immutability.

Two new instances based on the new singel-node overlay are added to the
dev environment under the human readable names of `arvo` and `mya`. The
rationale for naming is to make it easier to reference each of the
nodes in conversations. The nodes are configured to run on explicit AZs,
`us-east-2b` and `us-east-2c`, respectively, on `r5b.4xl` nodes. The new
nodes also use `30` bit bucket size in their storethehash datastore.
They are individually addressable via two separate ingress definitions.

A pod disruption budget is created which selects pods across both
deployments. This is to allow zero-downtime rollout of deployments
across both nodes, despite the fact that they are created by two
separate `Deployment` objects.

A new storage class of type `io2` with IOPS per GiB of 3 is created in
order to match the IOPS provided to production nodes in `dev`
environment. The new storage class is used by both of the new nodes.

Fixes: https://github.com/filecoin-project/storetheindex/issues/657