Deploy two individually configurable sti nodes in `prod`
Deploy two new individually configurable and publicly addressable
instances of storetheindex in `prod` with 30-bit sth bucket size running
on `r5b.4xl` workers.

Define a new storage class of type `io2` with 5 IOPS per GiB and use it
as the strorage class for the new nodes.

Both of the nodes are configured to run the head of `main` branch. Note
that CD pipeline is not enabled for the nodes, and change of version
running requires explicit PR to update the image tag in each relevant
node kustomization.