Roll out indexstar as the front-proxy in `prod`
Roll out the indexstar service as the front-proxy in production, to
scatter-gather requests across multiple indexer instances for the
following endpoints:
- `/reframe`
- `/multihash/*`
- `/cid/*`
- `/providers`

Scale up indexstar to 10 instances for high-availability and
load-handling. Deploy the latest version which includes a number of bug
fixes and improvements.

Enable CORS on indexstar ingress to keep web UI happy.