Do not delist blocked providers (#695)
* Do not delist blocked providers

When a provider is not allowed by policy, the indexer will not poll that provider and will remove it an delete its contents. This PR changes the behavior to retain blocked providers without removing them or their contents. This makes it safe to manipulate allow/block lists without the concern of deleting content.