Fix peering multiaddrs pointing to individual pods
Pods that are already in a service selector are individually addressable
within the same name space via `<pod-name>.<service-name>`. Update the
peering address to the indexer 0 and 1 accordingly.

While at it, peer nodes in `storetheindex-debug` with indexer 0 and 1 in
`storetheindex` namespace. Note that the dns name has to be FQDN since
the instances are in a different namespace.