Roll out GC free files option to dev and prod single instances
Roll out change to GC mechanism optionally freeing files in `dev` all
nodes, and `prod` single-deployment nodes. Preliminary testing showed
that changes to the locking mechanism significantly reduces spike in
latency when the GC runs. To further study the long-term impact of the
changes, the rollout here sets the option to `true` in some nodes and
`false` in others.