Experiment with an indexer instance with pebble store
Deploy a new node, called `oden` with identical infrastructure as
`vega`,
dedicated to ingest only from nft.storage, that uses pebble as its
valuestore.

Deploy `dido` same as `oden` but open to all providers to test how
pebble behaves without having to wait for the log chain to be downloaded
on oden. This would also give us a new datapoint on how pebble behaves
with multiple providers.

Relates to:
 - https://github.com/filecoin-project/go-indexer-core/pull/74