Keep entries chunk worker goroutines running
Do not start and stop entries chunk worker goroutines with each ad processed. Instead, if enabled, use a global pool that is at least as the worker pool size. The worker goroutines compete fairly for to use the entries chunk workers, so one will not starve others.

The config setting `Ingest.EntriesChunkConcurrency` controls the number of entries chunk workers and is runtime reloadable. A value of 0 (the default) sets this to the same as the ingest worker pool size.