Write entries chunks synchronously or asynchronously
There is no performance benefit to having a pool of goroutines over having one asynchronous goroutine per publisher/worker to asynchronously process and entries chunks. Having one additional goroutine provides a simpler implementation that uses less memory. It is also simpler to configure since the choice is to enable or disable the asynchronous behavior.

The default behavior is asynchronous.