Write entries chunks synchronously or asynchronously (#761)
* Write entries chunks synchronously or asynchronously

There is no performance benefit to having a pool of goroutines over having one asynchronous goroutine per publisher/worker to asynchronously process and entries chunks. Having one additional goroutine provides a simpler implementation that uses less memory. It is also simpler to configure since the choice is to enable or disable the asynchronous behavior. The behavior is configured with the reloadable config setting, `Ingest.WriteEntriesSynchronously`.

The default behavior is asynchronous.

* Make CorePutConcurrency run-time configurable
* STH gc resumes where it left off
* STH with open file cache
* log cache stats