Peer indexers to the two main instances (#711)
* Peer indexers to the two main instances

Enable ResendDirectAnnounce for the main indexer instances. This allows them to publish direct announce messages, received over to HTTP, to the other peer indexers via pubsub.

In the dev and prod cluster, peer the individual indexers to two main indexer nodes for the cluster.

* Fix peering multiaddrs pointing to individual pods

Pods that are already in a service selector are individually addressable
within the same namespace via `<pod-name>.<service-name>`. Update the
peering address to the indexer 0 and 1 accordingly.

While at it, peer nodes in `storetheindex-debug` with indexer 0 and 1 in
`storetheindex` namespace. Note that the DNS name has to be FQDN since
the instances are in a different namespace.

Co-authored-by: Masih H. Derkani <m@derkani.org>