Compare burst rate and sync interval
Do this on otherwise identical installation on romi and tara. It may not be sufficient to only increase sync interval with increasing burst rate as well, or vide versa, because one may become the limiting factor for the other. To see if these settings make a difference, increase them both on one node (tara).