Check if auto-sync is already running for a provider
If auto-sync is already running for a provider, do not run another one. If auto-sync is used to keep in sync, in the absence of announce messages, then it may be run on a small interval. This prevents auto-sync goroutines from building up if auto-sync is triggered while ingesting an advertisement chain.