Validate advertisement
This checks that the metadata and context ID is within the allowed limit, and if not causes the ad sync to fail. This was not being done previously and there is evidence that some oversized data was ingested on some indexers.

The sync is ended with an error, instead of skipping the advertisement, to limit the number of possibly huge unusable ads that are downloaded for ingestion.