Exclude `rami` from indexstar and adjust resources
`rami` is currently going through an index file upgrade; exclude it from
indexstar to avoid expensive timeouts until indexstar implements proper
exclusion of unhealthy backends.

The indexstar CPU usage was hovering at 70%; increase the CPU resource
to give it enough headroom at the face of request fluctuations. As a
result, up the max worker count in relevant EKS nodegroup to allow
scheduling of pods with higher CPU limit.

Reduce instance counts to 3 since most of the pods were mostly idle now
that resource limit is elevated.

Enable compression for `/reframe` endpoint on CloudFront; unrelated to
the timeout investigation. Added here since it was changed in the past
and terraform update was missing.

Relates to #863