Point landing page to `indexstar`
The landing page at `cid.contact` was powered by `indexer-0` and
`indexer-1`. Now that they are down it meant the landing page was
returning 503.

`indexstar` includes logic to bounce off landing page requests to the
first configured backend. Though not ideal this means the landing page
will continue to operate as long as the list of backends in `indexstar`
are maintained.

Note, this change means we cannot use `cid.contact` directly as a
backend in indexstar; otherwise it will result in an infinite redirect.