Point `dev.cid.contact` to CloudFront backed by `indexstar`
Change the routing of requests in `dev` environment, such that all
requests to `dev.cid.contact` go through CloudFront caching, and routed
via indexstar internally. This matches the routing mechanism currently
employed in `prod`.

As a result of changes, ingress to `indexer-0` and `indexer-1` is
changed to point to `indexer.dev.cid.contact` instead.

Fixes #856