Mark provider as Seen when sync is successful
A provider may be marked as inactive if its last contact time is not
updated for 2x longer than a configured threshold. The last contact time
was previously being updated only of there was some ads from a provider:
an active provider that has not published any new ads for a while could
get marked as inactive.

The changes here opportunistically marks a provider as seen: i.e.
whenever sync from that provider doesn't return an error. If subscriber.
Sync does not return an error it means there was some successful
communication to either fetch data or get the head.

Relates to #953