Include Bitswap providers with multiple transport in reframe response
The reframe server implementation only returned providers that supported
exactly one transport protocol in their metadata: `TransportBitswap`.

This implementation was added at the time where Provider record metadata
did not support multiple transports. Later, when the feature was added
in Metadata spec, the reframe server was not updated to reflect it.

The changes here fix the issue by checking if the metadata is prefixed
with `TransportBitswap` codec and if so it will include the records in
response. Check for prefix is sufficient here because, metadata bytes
must be sorted by codec as stated by the network indexer protocol.

Fixes #977