Upgrade CSI driver and install external snapshotter on `prod`
Upgrade to the latest CSI driver on `prod` now that it have been tested
on `dev` and deploy external snapshotter on `prod` so that we can take
snapshot of existing node's PVC and mount it as the state of new nodes.
This then would reduce time spent on catching up as well as making
testing on real data much easier.

Fixes #890
Fixes #891