Route HTTP announce requests to `kepa` for further propagation
Now that `indexer-1` is retried, and `indexer-0` is on its way to be no
longer, change the cloudfront routing to send all requests at `/ingest*`
to `kepa`. That node has the whitelisted peer ID on Lotus Bootstrap
nodes and is configured to propagate HTTP requests onto gossipsub
network for other indexer nodes. To reflect this change the node peering
configuration is also changed to connect to `kepa` instead of the old
instances.

The changes here also include increasing the max ASG for a node group
used to run indexer nodes just to have 1 node headroom, and adds missing
taints to `kepa` deployment object missed in earlier PR.

The work also removes a duplicate domain entry in ACM subject
alternative names to avoid redundant changes to the terraform state when
it is applied, since AWS removes the duplicates automatically and causes
the terraform state to be inconsistent with actual infrastructure state.