Track per-provider index counts and total (#984)
* Track per-provider index counts and total

Keep track of the number of indexes stored for each provider. For each provider, store a map of contextID to number of associated multihashes. When that context ID is removed, remove the entry from the map.  Get the total number of multihashes for a provider by summing all the counts for their contextIDs. Get the total stored index count by summing all counts for all contextIDs for all providers.

Counts are cached in memory after the first time they are retrieved.

Changes:
* Move index counters to separate package
* `/providers/{providerID}` API reports provider index count
* Configurable and reloadable IndexCountTotalAddend
* Unit and e2e tests for index counts
* IndexCounts works with both new sync and resync
* Show index count as part of provider info