Assigner service
Assigner service was previously called "Coordination service".

The Assigner service is a service implementation that is built and runs as a separate executable from storetheindex. The Assigner receives announce messages via pubsub and HTTP, and assigns the publishers described in these messages to the indexers configured as part of the Assigner's indexer pool.