Enable CD pipeline and retire `indexer-0` on `dev`
Enable CD pipeline for all `dev` nodes, such that PRs opened by CD as a
result of merges to `main` will deploy the latest to all `dev` nodes
(currently two `ber` and `cali`). This is done by removing `image`
overrides in `instances/<node-name>/kustomization.yaml`; which means if
we need to override the CD on any one of the nodes and deploy a specific
version all that's needed is to add back the `image` override in the
corresponding node's `kustomization.yaml` file.

Retire `indexer-0` the only remaining sth-backed node on dev and route
all `/ingest/*` requests to `ber` since it is using an identity
whitelisted by Lotus bootstrap nodes and has re-propagation of HTTP
announces over gossipsub (i.e. `ResendDirectAnnounce` enabled.

While at it, move GitHub authorisation for storetheindex tenant to the
cluster level kustomizations to reduce irrelevant manifests in
directories that are meant to contain application-level stuff. The
secret is re-encrypted with the cluster level key.